+++ 成冬樊task01/Staff/charge/Charge.java
@@ -1,0 +1,25 @@
+﻿package Staff.charge;
+import java.util.Scanner;
+import Staff.member.Member;
+/**
+ *
+ * @author chengdongfan;
+ * @version V1.0
+ * @date 2018/10/24
+ * @since 1.8
+ */
+public class Charge extends Member {
+    Scanner inputData = new Scanner(System.in);
+
+    //工资范围
+    public void setSalary(double salary) {
+        if (salary > 6000 && salary < 8000) {
+            super.setSalary(salary);
+        } else {
+            System.out.println("输入薪水有误,请重新输入薪水:");
+            salary = inputData.nextInt();
+            super.setSalary(salary);
+        }
+
+    }
+}
