--- 成冬樊task01/team/employee/charge/Charge.java
@@ -1,16 +1,0 @@
-package team.employee.charge;
-
-import team.employee.member.Member;
-
-public class Charge extends Member {
-
-     //工资范围
-    public void setSalary(double salary) {
-        if (salary > 6000 && salary < 8000) {
-            super.setSalary(salary);
-        }
-
-    }
-
-}
-
